adapt javadocs
