+++ src/test/java/de/muenchen/isi/infrastructure/repository/OptimisticLockingTest.java
@@ -20,11 +20,7 @@ import org.springframework.test.context.ActiveProfiles;
 
 @Slf4j
 @ExtendWith(MockitoExtension.class)
-@SpringBootTest(
-    classes = { IsiBackendApplication.class },
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-    properties = { "tomcat.gracefulshutdown.pre-wait-seconds=0" }
-)
+@SpringBootTest(classes = { IsiBackendApplication.class }, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ActiveProfiles(profiles = { TestConstants.SPRING_UNIT_TEST_PROFILE, TestConstants.SPRING_NO_SECURITY_PROFILE })
 @MockitoSettings(strictness = Strictness.LENIENT)
 public class OptimisticLockingTest {
